Split namespaces utils into multiple files
